+++ plugins/module_utils/netbox_utils.py
@@ -12,7 +12,6 @@ __metaclass__ = type
 import traceback
 import re
 import json
-
 from itertools import chain
 
 from ansible.module_utils.common.text.converters import to_text
@@ -70,7 +69,7 @@ API_APPS_ENDPOINTS = dict(
         "site_groups",
         "virtual_chassis",
     ],
-    extras=["tags"],
+    extras=["config_contexts", "tags"],
     ipam=[
         "aggregates",
         "ip_addresses",
@@ -96,6 +95,7 @@ QUERY_TYPES = dict(
     cluster="name",
     cluster_group="slug",
     cluster_type="slug",
+    config_context="name",
     device="name",
     device_role="slug",
     device_type="slug",
@@ -107,7 +107,9 @@ QUERY_TYPES = dict(
     manufacturer="slug",
     nat_inside="address",
     nat_outside="address",
+    parent_location="slug",
     parent_region="slug",
+    parent_site_group="slug",
     parent_tenant_group="slug",
     power_panel="name",
     power_port="name",
@@ -150,7 +152,9 @@ CONVERT_TO_ID = {
     "circuits.circuittermination": "circuit_terminations",
     "cluster": "clusters",
     "cluster_group": "cluster_groups",
+    "cluster_groups": "cluster_groups",
     "cluster_type": "cluster_types",
+    "config_context": "config_contexts",
     "dcim.consoleport": "console_ports",
     "dcim.consoleserverport": "console_server_ports",
     "dcim.frontport": "front_ports",
@@ -162,6 +166,7 @@ CONVERT_TO_ID = {
     "device": "devices",
     "device_role": "device_roles",
     "device_type": "device_types",
+    "device_types": "device_types",
     "export_targets": "route_targets",
     "group": "tenant_groups",
     "import_targets": "route_targets",
@@ -177,10 +182,13 @@ CONVERT_TO_ID = {
     "nat_inside": "ip_addresses",
     "nat_outside": "ip_addresses",
     "platform": "platforms",
+    "parent_location": "locations",
     "parent_interface": "interfaces",
     "parent_vm_interface": "interfaces",
     "parent_region": "regions",
+    "parent_site_group": "site_groups",
     "parent_tenant_group": "tenant_groups",
+    "platforms": "platforms",
     "power_panel": "power_panels",
     "power_port": "power_ports",
     "power_port_template": "power_port_templates",
@@ -193,19 +201,24 @@ CONVERT_TO_ID = {
     "rack_group": "rack_groups",
     "rack_role": "rack_roles",
     "region": "regions",
+    "regions": "regions",
     "rear_port": "rear_ports",
     "rear_port_template": "rear_port_templates",
     "rir": "rirs",
+    "roles": "device_roles",
     "route_targets": "route_targets",
     # Just a placeholder as scope can be several different types including sites.
     "scope": "sites",
     "services": "services",
     "site": "sites",
     "site_group": "site_groups",
+    "sites": "sites",
     "tags": "tags",
     "tagged_vlans": "vlans",
     "tenant": "tenants",
+    "tenants": "tenants",
     "tenant_group": "tenant_groups",
+    "tenant_groups": "tenant_groups",
     "termination_a": "interfaces",
     "termination_b": "interfaces",
     "untagged_vlan": "vlans",
@@ -227,6 +240,7 @@ ENDPOINT_NAME_MAPPING = {
     "clusters": "cluster",
     "cluster_groups": "cluster_group",
     "cluster_types": "cluster_type",
+    "config_contexts": "config_context",
     "console_ports": "console_port",
     "console_port_templates": "console_port_template",
     "console_server_ports": "console_server_port",
@@ -285,6 +299,21 @@ ALLOWED_QUERY_PARAMS = {
     "cluster": set(["name", "type"]),
     "cluster_group": set(["slug"]),
     "cluster_type": set(["slug"]),
+    "config_context": set(
+        [
+            "name",
+            "regions",
+            "sites",
+            "roles",
+            "device_types",
+            "platforms",
+            "cluster_groups",
+            "clusters",
+            "tenant_groups",
+            "tenants",
+            "tags",
+        ]
+    ),
     "console_port": set(["name", "device"]),
     "console_port_template": set(["name", "device_type"]),
     "console_server_port": set(["name", "device"]),
@@ -316,9 +345,11 @@ ALLOWED_QUERY_PARAMS = {
     "manufacturer": set(["slug"]),
     "master": set(["name"]),
     "nat_inside": set(["vrf", "address"]),
+    "parent_location": set(["slug"]),
     "parent_interface": set(["name"]),
     "parent_vm_interface": set(["name"]),
     "parent_region": set(["slug"]),
+    "parent_site_group": set(["slug"]),
     "parent_tenant_group": set(["slug"]),
     "platform": set(["slug"]),
     "power_feed": set(["name", "power_panel"]),
@@ -411,9 +442,11 @@ CONVERT_KEYS = {
     "circuit_type": "type",
     "cluster_type": "type",
     "cluster_group": "group",
+    "parent_location": "parent",
     "parent_interface": "parent",
     "parent_vm_interface": "parent",
     "parent_region": "parent",
+    "parent_site_group": "parent",
     "parent_tenant_group": "parent",
     "power_port_template": "power_port",
     "prefix_role": "role",
@@ -508,6 +541,7 @@ class NetboxModule(object):
             self.nb = nb_client
             try:
                 self.version = self.nb.version
+                self.full_version = self.nb.status().get("netbox-version")
             except AttributeError:
                 self.module.fail_json(msg="Must have pynetbox >=4.1.0")
 
@@ -556,6 +590,7 @@ class NetboxModule(object):
             nb.http_session = session
             try:
                 self.version = nb.version
+                self.full_version = nb.status().get("netbox-version")
             except AttributeError:
                 self.module.fail_json(msg="Must have pynetbox >=4.1.0")
             except Exception:
@@ -907,7 +942,7 @@ class NetboxModule(object):
                         self.version, "2.9", greater_or_equal=True
                     )
                     and k == "tags"
-                ):
+                ) or (self.endpoint == "config_contexts" and k == "tags"):
                     continue
                 if k == "termination_a":
                     endpoint = CONVERT_TO_ID[data.get("termination_a_type")]
@@ -936,7 +971,18 @@ class NetboxModule(object):
                 elif isinstance(v, list):
                     id_list = list()
                     for list_item in v:
-                        if k == "tags" and isinstance(list_item, str):
+                        if k in (
+                            "regions",
+                            "sites",
+                            "roles",
+                            "device_types",
+                            "platforms",
+                            "cluster_groups",
+                            "clusters",
+                            "tenant_groups",
+                            "tenants",
+                            "tags",
+                        ) and isinstance(list_item, str):
                             temp_dict = {"slug": self._to_slug(list_item)}
                         elif isinstance(list_item, dict):
                             norm_data = self._normalize_data(list_item)
@@ -1097,6 +1143,21 @@ class NetboxModule(object):
         if serialized_nb_obj.get("tags") and data.get("tags"):
             serialized_nb_obj["tags"] = set(serialized_nb_obj["tags"])
             updated_obj["tags"] = set(data["tags"])
+
+        # Ensure idempotency for site on version pre-3.0
+        version_pre_30 = self._version_check_greater("3.0", self.version)
+        if (
+            serialized_nb_obj.get("latitude")
+            and data.get("latitude")
+            and version_pre_30
+        ):
+            updated_obj["latitude"] = str(data["latitude"])
+        if (
+            serialized_nb_obj.get("longitude")
+            and data.get("longitude")
+            and version_pre_30
+        ):
+            updated_obj["longitude"] = str(data["longitude"])
 
         if serialized_nb_obj == updated_obj:
             return serialized_nb_obj, None