+++ plugins/modules/netbox_site.py
@@ -69,6 +69,7 @@ options:
           - The site group the site will be associated with (NetBox 2.11+)
         required: false
         type: raw
+        version_added: "3.3.0"
       tenant:
         description:
           - The tenant the site will be assigned to
@@ -108,12 +109,12 @@ options:
         description:
           - Latitude in decimal format
         required: false
-        type: str
+        type: float
       longitude:
         description:
           - Longitude in decimal format
         required: false
-        type: str
+        type: float
       contact_name:
         description:
           - Name of contact for site
@@ -266,8 +267,8 @@ def main():
                     description=dict(required=False, type="str"),
                     physical_address=dict(required=False, type="str"),
                     shipping_address=dict(required=False, type="str"),
-                    latitude=dict(required=False, type="str"),
-                    longitude=dict(required=False, type="str"),
+                    latitude=dict(required=False, type="float"),
+                    longitude=dict(required=False, type="float"),
                     contact_name=dict(required=False, type="str"),
                     contact_phone=dict(required=False, type="str"),
                     contact_email=dict(required=False, type="str"),